awscdk-minecraft/pyproject.toml
[project]
name = "awscdk-minecraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib==2.154.1",
    "boto3>=1.35.44",
    "constructs>=10.0.0,<11.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cdk_minecraft"]
